+++ packages/locale/utils/create_packages.js
@@ -32,6 +32,9 @@ const yargs = yargsPkg
             description: 'output debug infos'
         },
     })
+    .parserConfiguration({
+        'camel-case-expansion': false
+    })
     .wrap(Math.min(120, yargsPkg.terminalWidth()))
     .help();
 
@@ -53,6 +56,7 @@ const packageTemplate = {
         url: 'https://github.com/js-joda/js-joda-locale.git'
     },
     main: 'dist/index.js',
+    typings: 'dist/js-joda-locale.d.ts',
     keywords: [
         'date',
         'time',
@@ -76,7 +80,8 @@ const packageTemplate = {
     devDependencies: {}
 };
 
-const readmeTemplate = fs.readFileSync(path.resolve(__dirname, 'README_package.template.md'), 'utf8');
+const readmeTemplate = fs.readFileSync(path.resolve(__dirname, 'README_package.template.md'),
+    'utf8');
 const readmeLocaleRegex = /{{locale}}/g;
 
 // TODO: build minified (DIST_MIN=1) package?? in dist/index.min.js?
@@ -87,6 +92,10 @@ Object.keys(argv.packages).forEach((packageName) => {
     if (!fs.existsSync(packageDir)) {
         fs.mkdirSync(packageDir);
     }
+    const distDir = path.resolve(argv.packagesDir, packageName, 'dist');
+    if (!fs.existsSync(distDir)) {
+        fs.mkdirSync(distDir);
+    }
     // create package.json
     packageTemplate.version = mainPackageJSON.version;
     packageTemplate.name = `@js-joda/locale_${packageName}`;
@@ -95,6 +104,8 @@ Object.keys(argv.packages).forEach((packageName) => {
         JSON.stringify(packageTemplate, null, 4));
     fs.writeFileSync(path.resolve(packageDir, 'README.md'),
         readmeTemplate.replace(readmeLocaleRegex, argv.packages[packageName].join(',')));
+    fs.copyFileSync(path.resolve(__dirname, '..', 'dist', 'js-joda-locale.d.ts'),
+        path.resolve(packageDir, 'dist', 'js-joda-locale.d.ts'));
     const nodeArgs = [
         './utils/build_package.js',
         '-o', `${path.resolve(packageDir, 'dist')}`,