+++ packages/timezone/CHANGELOG.md
@@ -3,6 +3,14 @@ Changelog
 
 ### next
 
+### 2.5.0
+
+ * updated to iana tzdb 2020d 
+ 
+### 2.4.0
+
+ * updated to iana tzdb 2020c 
+ 
 ### 2.3.0
 
  * updated to iana tzdb 2020a